# Optimize findAccessKey