With this change, we spend half the CPU and allocate 1/20 of the memory, with half the number of allocations. Note how TCP and UDP are quite on par after the change.

Before:
```
BenchmarkTCPFindCipher-12    	    1000	   2441318 ns/op	 2026527 B/op	    3013 allocs/op
BenchmarkUDPUnpack-12        	    2000	    919961 ns/op	  125024 B/op	    1701 allocs/op
```

After:
```
BenchmarkTCPFindCipher-12    	    1000	   1353308 ns/op	  125268 B/op	    1705 allocs/op
BenchmarkUDPUnpack-12        	    2000	    890936 ns/op	  125029 B/op	    1701 allocs/op
```

 The optimization required dealing with low level parts of the protocol. I'm happy to explain it in person.